upload 2 cf problem & modified 1

+++ codeforces/assiut_university_training-newcomers/sheet_06_math-geometry/T_Circle_Task.cpp
@@ -9,12 +9,12 @@ URL: https://codeforces.com/group/MWSDmqGsZm/contest/223338/problem/T
 #define nl '\n'
 using namespace std;
 
-ll distance(ll x, ll y, ll x1, ll y1)
+double distance(ll x, ll y, ll x1, ll y1)
 {
     ll a = pow(x1-x, 2);
     ll b = pow(y1-y, 2);
-    ll res = sqrt(a+b);
-
+    double res = sqrt(a+b);
+    
     return res;
 }
 
@@ -28,7 +28,7 @@ int main(){
     {
         ll x1, y1; cin >> x1 >> y1;
         
-        ll dis = distance(x, y, x1, y1);
+        double dis = distance(x, y, x1, y1);
         if(dis <= radius) cout << "YES" << nl;
         else cout << "NO" << nl;
